Query: Speed improvement for scalar elements.
Now the components array is created in the constructor and stored
as a field. This doubles the performance of code that likes to
view all elements as vectors (e.g. the statistics-gathering code).
See benchmark tests in `ElementSpeedTest`.

This seems to have negligible effect on other queries. The
benchmark results in `benchmarks_GPU-Machine.xml` have become
slightly worse, it's true - however even without this change, the
results are similar. I.e. the slowdown is due to something else;
either an earlier code change or a system change on the computer
GPU-Machine.