Query: Removed prototype elements from statistics classes.
Because they are hard to store in the database. So, we use primitive
types instead e.g. in the Cats class for the key. This means we
have to choose what kind of key to use; we have chosen ints. Cats
can now only be used with metadata bands of type byte, short and int.
... It's kind of silly to use longs, doubles and floats for categories
anyway.