Fixed up Tile Map Server servlet
Renamed from wmts to tms as it is a Tile Map Sever and not a Web Map Tile Service (they are different). Url now works much better with a new format, better support for query requests (although still not implemented), and will now look up timeslice (latest) if one is not provided.