Moved config to a shared location
The storage manager tests were trying to load resources that
weren't in the classpath. The paths have now been fixed. As part
of that, the config directory has been moved to a shared location,
and the SAMPLE suffix has been removed in favour of providing a
Docker volume override.

The default database credentials have been changed to match those
provided by _/postgres:9 on Docker Hub.