Generating command line scripts for cmdclient and rsaquery
This functionality used to be provided by a hand-written script,
cmdclinet/scripts/rsa. But Gradle has an application plugin that
does pretty much the same thing. Now, if you run:

    cd src/cmdclient
    gradle installDist

Then you'll get a script that you can run in:

    build/install/cmdclient/bin/cmdclient

And if you run:

    gradle distZip

Then you'll get a zip file containing the script and all required
libraries in:

    build/distributions/cmdclient*.zip.