Using Gradle Tomcat plugin for web services
Added rules to spatialcubeservice/build.gradle to install and use
an embedded Tomcat server. But currently it doesn't properly
start: no errors, but the web server doesn't seem to be listening.

To get this working to the point that it is, the libraries in
WEB-INF/lib needed to be removed - because some of them conflict
with the libraries provided by Tomcat itself.