Automatically setting seed IP in config file
When starting an Akka seed node, the node's IP address will be
inserted into the config file if the config file doesn't yet
contain that information. This makes it easier to launch the akka
system in a development environment. If the file already contains
the seed IP addresses, nothing happens (for deployment).