Including config directory in zip files when creating distribution
A link to the config directory is included in src/dist of cmdclient
and rsaworkers. This is then included in the zip file when running
`gradle distZip`. The directory is overridden in `installDist` by
an explicit delete command, and it is then replaced by a symlink.

Also cleaned up the .gitignore files, that contained a lot of
reduntant information, an which were making it hard to see notice
new files under src/dist.