Upgraded to Hibernate 4.2 and Spring 3.2
This seems to be required to get array IO working.

The construction of DAOs has changed: there is no more
HibernateTempalte, so the SessionFactor is used instead. The
Session no longer has a find() method, so createQuery().list()
is used.

EnumType has been removed, because there's a built-in Hibernate
class that does pretty much the same thing.