Upgrade to Spring 4
Upgraded Spring to version 4.2. A number of classes needed to
change imports, because packages have changed. Some peripheral
libraries needed to be upgraded too, e.g. Jackson.

The configuration of Spring MVC has been updated to work with
the new version. It now uses a ContentNegotiationManager instead
of just a ContentNegotiatingViewResolver. It no longer lists
all classes that can be served via XML; instead, it specifies
the package that they all belong to.

Also fixed serialisation of TabularResponse: it was incorrectly
declaring some of its elements as attributes.