Fixed mapping of embedded lists
The remining issues were mostly to do with the actual mapping file:
the List<String> was not declaring itself as such, so it was trying
to store as a List<Double>. And some other minor errors.

Added a logging instruction for Postgres. This is
counter-intuitively set in the database connection string, not in
the logging config file.