Using a simplified table format for tabular data
Previously, table cells were accessed by name, e.g. `row['area']`.
This was fine when tables only had a fixed number of columns - but
a Ledger can have any number of columns. The TabularResponse now
contains a 2D array of cells of numbers, which must be accessed by
index.

Currently, this is working well for JSON, but not XML: in XML, the
rows are empty.