Improved integration with Landblade
Changed the default config directory to ../config/rsa instead of
./config. This means it's easier to override the config without
git noticing.

Also, running Gradle before the source is available to download
dependencies. This makes rebuilds faster, because there is a Docker
checkpoint after the initial build. Gradle is run again later in
the build process, after the source is copied in to the image.