Simplified docker-compose deployment: using seed hostname
Previously, the seed node would modify its config file on startup
to insert its own IP address. This was unreliable, because if the
container didn't exit properly then the next time a seed started
it would not modify the file.

Now, the config file just uses the "seed" hostname, which is set
by docker-compose for the appropriate container. The workers
are aware of it via a container link.