Fixed seed config: finding out IP address
Akka really really likes to be configured by IP address instead of
by hostname. But for docker-compose, it's much easier to specify
the hostname. Now the hostname is converted to an IP address during
initialisation. Without this, the seed node complains that the
messages it is receiving are meant for someone else.