Accountant filter now ignores null values
If all inputs are NODATA for any pixel, then that pixel will not
be stored in the ledger at all. If some pixels are real data and
some are NODATA, the NODATA values are stored in the ledger as
null.

When fetching a table from the web services, if a ledger has null
rows after filtering, those rows are excluded from the results.