Fixed cyclic dependency in dev service
The dev service was depending on the seed service, but the seed
service couldn't be started until gradle had been run inside the
dev service. The fix was to migrate to version 2 of the
docker-compose config format. We're now using depends_on instead
of links to control startup. The dev container no longer depends
on anything, but - because of version 2's networking system -
you can still access the other services if they happen to be
running.