disable passing zoom levels to gdal2tiles by default
Default behaviour of gdal2tiles is to best match the highest resolution to the source data. Overriding this can result in generating more levels of map tiles than is necessary. In future it may be required batch requests to gdal2tiles, with different batched for different zoom levels, it is for this reason the option to specify zoom levels remains.