Fixed copy task
When the config directory is mounted read-only, it was possible
to get into a state when the installDist task would fail when
copying the config. The solution is to use two delete instructions:
one before and one after the main installDist task. Then the
symlink to the config directory can be created.